{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = ["loguru", "python-dotenv"]

[project.optional-dependencies]
dev = ["pytest", "pre-commit", "ruff", "ty"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/{{ cookiecutter.project_slug }}"]

# ===== RUFF =====
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "D",    # pydocstyle
    "ARG",  # flake8-unused-arguments
    "RET",  # flake8-return
    "C4",   # flake8-comprehensions
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]

# ===== TY =====
[tool.ty]
python-version = "3.12"
